requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
courlan==0.9.2
dataclasses-json==0.5.7
dateparser==1.1.8
discord.py==2.2.3
frozenlist==1.3.3
greenlet==2.0.2
htmldate==1.4.3
idna==3.4
jusText==3.0.0
langchain==0.0.175
langcodes==3.3.0
lxml==4.9.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.7
openapi-schema-pydantic==1.2.4
packaging==23.1
pydantic==1.10.7
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
six==1.16.0
SQLAlchemy==2.0.15
tenacity==8.2.2
tld==0.13
tqdm==4.65.0
trafilatura==1.6.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.2
yarl==1.9.2
